test support refactor
