+++ tsconfig.json
@@ -33,7 +33,8 @@
     "paths": {
       "@lib/*": [ "src/*" ],
       "@backend/*": ["src/queue/backend/*"],
-      "@steps/*": [ "features.steps/*" ]
+      "@steps/*": [ "features.steps/*" ],
+      "@test": ["src/utils/testSupport.ts"]
     },                          
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     "types": ["node", "mocha"],
